refactor: split upload file to helper and pipeline
